Renaming PipelineResourcesHolder.h/cpp to QueryPlanResourceHolder.h/cpp
