+++ src/QueryPipeline/QueryPlanResourceHolder.h
